refactor: bytebase-action no global vars
preparing for testing
