+++ action/world/world.go
@@ -1,0 +1,43 @@
+package world
+
+// World is the world environment for bytebase-action.
+type World struct {
+	// bytebase-action flags
+	Output               string
+	URL                  string
+	ServiceAccount       string
+	ServiceAccountSecret string
+	Project              string // projects/{project}
+	Targets              []string
+	FilePattern          string
+
+	// bytebase-action check flags
+	// An enum to determine should we fail on warning or error.
+	// Valid values:
+	// - SKIP
+	// - FAIL_ON_WARNING
+	// - FAIL_ON_ERROR
+	CheckRelease string
+
+	// bytebase-action rollout flags
+	ReleaseTitle string // The title of the release
+	RolloutTitle string // The title of the rollout
+	// An enum to determine should we run plan checks and fail on warning or error.
+	// Valid values:
+	// - SKIP
+	// - FAIL_ON_WARNING
+	// - FAIL_ON_ERROR
+	CheckPlan string
+	// Rollout up to the target-stage.
+	// Format: environments/{environment}
+	TargetStage string
+	Plan        string
+
+	OutputMap map[string]string
+}
+
+func NewWorld() *World {
+	return &World{
+		OutputMap: make(map[string]string),
+	}
+}
